added a mock application class
